requirements.txt
streamlit
google-generativeai
spotipy